Manage installation of the external project manually.
This change, mostly influenced by https://github.com/ament/uncrustify_vendor/,
modifies the cmake arguments for tinydir installation to install it into
the build directory of the vendor package, then the vendor package takes
that result and installs it to the ultimate target during the vendor
package's install phase.

This fixes #2.